Messy, but wanted to note confusion...
I got myself confused trying to add another event listener/handler
refactoring the move list into components so that I can bold one of
them.

I'm not sure this was the mistake I started with, but ultimately my
mistake was trying to pass in an arg in the beginning of the arrow fn
(x) => onFoo(x) which passes in some context?

What I want to do: on the lowest down element (the one with access to
the data I want), have no arg on the left and pass in the arg you want
on the right. on higher-up onclick attrs go ahead and put (x) =>
onFoo(x) in order to keep passing the data up to the handler.